Check uniqueness of a new patient's friendly ID in the controller
