+++ app/models/abstract.js
@@ -3,10 +3,8 @@ import Ember from 'ember';
 import EmberValidations from 'ember-validations';
 import { Model } from 'ember-pouch';
 import UserSession from 'hospitalrun/mixins/user-session';
-import PouchDbMixin from 'hospitalrun/mixins/pouchdb';
-export default Model.extend(UserSession, EmberValidations, PouchDbMixin, {
+export default Model.extend(UserSession, EmberValidations, {
   session: Ember.inject.service(),
-  database: Ember.inject.service(),
   archived: DS.attr('boolean'),
   lastModified: DS.attr('date'),
   modifiedBy: DS.attr(),
