+++ app/patients/edit/controller.js
@@ -7,8 +7,6 @@ import ReturnTo from 'hospitalrun/mixins/return-to';
 import SelectValues from 'hospitalrun/utils/select-values';
 import UserSession from 'hospitalrun/mixins/user-session';
 export default AbstractEditController.extend(BloodTypes, ReturnTo, UserSession, PatientId, PatientNotes, {
-  config: Ember.inject.service(),
-
   canAddAppointment: function() {
     return this.currentUserCan('add_appointment');
   }.property(),
@@ -79,6 +77,7 @@ export default AbstractEditController.extend(BloodTypes, ReturnTo, UserSession, 
     'Private'
   ],
 
+  config: Ember.inject.service(),
   filesystem: Ember.inject.service(),
   database: Ember.inject.service(),
   patientController: Ember.inject.controller('patients'),
@@ -549,11 +548,27 @@ export default AbstractEditController.extend(BloodTypes, ReturnTo, UserSession, 
   },
 
   beforeUpdate: function() {
-    return this.get('model').validate().catch((err) => {
-      return this.generateFriendlyId().then((friendlyId) => {
-        this.model.set('friendlyId', friendlyId);
+    if (!this.get('model.isNew')) {
+      return Ember.RSVP.resolve();
+    }
+    const database = this.get('database');
+    const id = this.get('model.friendlyId');
+    const maxValue = this.get('maxValue');
+    const query = {
+      startkey: [id, null],
+      endkey: [id, maxValue]
+    };
+    return database.queryMainDB(query, 'patient_by_display_id')
+      .then((found) => {
+        if (Ember.isEmpty(found.rows)) {
+          return Ember.RSVP.resolve();
+        }
+        return this.generateFriendlyId()
+          .then((friendlyId) => {
+            this.model.set('friendlyId', friendlyId);
+            return Ember.RSVP.resolve();
+          });
       });
-    });
   },
 
   afterUpdate: function(record) {
